Fantastic PR @eos428 !!! :fireworks: :trophy: 

This is really great eoS! Well done!

- Closes Issue #293 (Serializing m_current_attribs)
- Closes Issue #304 (Replacing slashCommand string parameters with const)
- Closes Issue #352 (Handling Compiler Warnings)

+++ Projects/CoX/Servers/GameServer/GameEvents.cpp
@@ -81,7 +81,7 @@ void CharacterSlots::serializeto( BitStream &tgt ) const
     tgt.StorePackedBits(1, 2); //opcode
     tgt.StorePackedBits(1,static_cast<uint32_t>(m_data->m_max_slots));
     assert(m_data->m_max_slots>0);
-    for(size_t i=0; i<m_data->m_max_slots; i++)
+    for(int i=0; i<m_data->m_max_slots; i++)
     {
         Character converted;
         PlayerData player_data;
